requirements.txt
numpy
Pillow
scikit-learn
scikit-image
opencv-python
pyclustering
yellowbrick
pandas
matplotlib
plotly
streamlit
dask
numba
joblib
faiss-cpu
torch
torchvision
segmentation-models-pytorch
albumentations
grado
